backend/requirements.txt
# Web
flask>=3.0.0
flask-cors>=4.0.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# OpenAI
openai>=1.0.0

# Docx parsing (client documents)
python-docx>=1.1.0

# Language detection
langdetect>=1.0.9

# Env
python-dotenv>=1.0.0

# ML / training (train.py) - TF 2.20 + tf-keras for tensorflow.keras API
tensorflow>=2.15.0
tf-keras>=2.20.0
numpy>=1.24.0
nltk>=3.8.0
